use ffmpeg-sys

+++ src/error.rs
@@ -4,7 +4,7 @@ macro_rules! check_result {
     let errnum = $condition;
     if errnum < 0 {
       let mut data = [0i8; AV_ERROR_MAX_STRING_SIZE];
-      av_strerror(errnum, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE as u64);
+      av_strerror(errnum, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE);
       $block;
       return Err(tools::to_string(data.as_ptr()));
     }
@@ -13,7 +13,7 @@ macro_rules! check_result {
     let errnum = $condition;
     if errnum < 0 {
       let mut data = [0i8; AV_ERROR_MAX_STRING_SIZE];
-      av_strerror(errnum, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE as u64);
+      av_strerror(errnum, data.as_mut_ptr(), AV_ERROR_MAX_STRING_SIZE);
       return Err(tools::to_string(data.as_ptr()));
     }
   };
